# docs/swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
  dto.LoginRequest:
    properties:
      email:
        example: me@tfkhdyt.my.id
        type: string
      password:
        example: bruh1234
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginResponseData'
      message:
        example: You've logged in successfully
        type: string
    type: object
  dto.LoginResponseData:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  dto.LogoutResponse:
    properties:
      message:
        example: You've logged out successfully
        type: string
    type: object
  dto.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/dto.RefreshResponseData'
      message:
        example: Your access token has been refreshed successfully
        type: string
    type: object
  dto.RefreshResponseData:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: me@tfkhdyt.my.id
        type: string
      full_name:
        example: Taufik Hidayat
        maxLength: 70
        minLength: 2
        type: string
      password:
        example: bruh1234
        minLength: 8
        type: string
      username:
        example: tfkhdyt
        maxLength: 16
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.RegisterResponse:
    properties:
      data:
        $ref: '#/definitions/dto.RegisterResponseData'
      message:
        example: Your account has been created successfully
        type: string
    type: object
  dto.RegisterResponseData:
    properties:
      created_at:
        example: "2023-09-04T21:00:43.775157Z"
        type: string
      email:
        example: me@tfkhdyt.my.id
        type: string
      full_name:
        example: Taufik Hidayat
        type: string
      id:
        example: 69
        type: integer
      username:
        example: tfkhdyt
        type: string
    type: object
  exception.HttpError:
    properties:
      error:
        example: Something went wrong
        type: string
    type: object
  exception.ValErrors:
    properties:
      errors:
        example:
        - Invalid email
        - Username is required
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: me@tfkhdyt.my.id
    name: API Support
    url: https://tfkhdyt.my.id
  description: SpaceNotes REST API server
  license:
    name: MIT License
    url: https://github.com/spacenotes-id/server/blob/main/LICENSE
  title: SpaceNotes API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to get access token and refresh token
      parameters:
      - description: Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/exception.ValErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpError'
      summary: Login
      tags:
      - auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Remove refresh token from database
      parameters:
      - description: Refresh token
        in: body
        name: refreshToken
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LogoutResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exception.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/exception.ValErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpError'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /auth/refresh:
    patch:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refreshToken
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exception.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/exception.ValErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpError'
      summary: Refresh
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new account
      parameters:
      - description: Request body
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/exception.ValErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpError'
      summary: Register
      tags:
      - auth
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    description: JWT key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"